+++ src/main/java/lumien/randomthings/entitys/EntitySoul.java
@@ -1,14 +1,10 @@
 package lumien.randomthings.entitys;
 
-import java.util.List;
-
 import io.netty.buffer.ByteBuf;
-import lumien.randomthings.item.ItemRezStone;
 import lumien.randomthings.util.PlayerUtil;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.AxisAlignedBB;
@@ -82,7 +78,7 @@ public class EntitySoul extends Entity implements IEntityAdditionalSpawnData
 	@Override
 	public boolean canBeCollidedWith()
 	{
-		return true;
+		return false;
 	}
 
 	@Override
@@ -94,23 +90,6 @@ public class EntitySoul extends Entity implements IEntityAdditionalSpawnData
 	@Override
 	public final boolean processInitialInteract(EntityPlayer player, EnumHand hand)
 	{
-		ItemStack equipped = player.getHeldItemMainhand();
-		if (!world.isRemote && !equipped.isEmpty() && equipped.getItem() instanceof ItemRezStone && this.world.getMinecraftServer().getPlayerList().getPlayerByUsername(playerName) != null)
-		{
-			List<EntityReviveCircle> circles = this.world.getEntitiesWithinAABB(EntityReviveCircle.class, new AxisAlignedBB(this.posX - 2, this.posY - 2, this.posZ - 2, this.posX + 2, this.posZ + 2, this.posZ + 2));
-
-			for (EntityReviveCircle circle : circles)
-			{
-				if (circle.toRevive == this)
-				{
-					return false;
-				}
-			}
-
-			world.spawnEntity(new EntityReviveCircle(world, player, posX, posY, posZ, this));
-
-			return true;
-		}
 		return false;
 	}